apps/local-oidc-mock/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
PyJWT[crypto]>=2.10.1,<3
cryptography>=44.0.0
